Move umbrella event checkboxes to type section
